+++ Runtime/NativePriorityQueueDispose.cs
@@ -1,16 +1,18 @@
-﻿using Unity.Collections;
+﻿using System;
+using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
 
 namespace BonnFireGames.CustomNativeContainers
 {
     [NativeContainer]
-    public struct NativePriorityQueueDispose
+    [BurstCompatible(GenericTypeArguments = new [] { typeof(int) })]
+    internal struct NativePriorityQueueDispose<T> where T : unmanaged, IComparable<T>
     {
         [NativeDisableUnsafePtrRestriction]
-        internal unsafe void* m_Buffer;
-        internal Allocator m_AllocatorLabel;
+        internal unsafe NativePriorityQueueData<T>* Data;
+        internal AllocatorManager.AllocatorHandle AllocatorHandle;
         internal AtomicSafetyHandle m_Safety;
 
-        public unsafe void Dispose() => UnsafeUtility.FreeTracked(this.m_Buffer, this.m_AllocatorLabel);
+        public unsafe void Dispose() => NativePriorityQueueData<T>.DeallocateQueue(Data, AllocatorHandle);
     }
 }